requirements.txt
# pip install -r requirements.txt

torch>=1.10.2
torchvision>=0.11.3
numpy~=1.18.5
pandas>=1.3.5
jupyter

PyQt5>=5.15.6

kiwisolver>=1.3.2
pillow~=9.0.0
opencv-python~=4.5.5.62
matplotlib~=3.5.1